Move routing to package
