+++ packages/equipment-experts/package.json
@@ -5,12 +5,14 @@
   "repository": "https://github.com/parameter1/randall-reilly-websites/tree/master/packages/equipment-experts",
   "license": "MIT",
   "private": true,
+  "main": "index.js",
   "scripts": {
     "lint": "eslint --ext .js --ext .vue --max-warnings 5 --ignore-path ../../.eslintignore ./",
     "test": "yarn lint"
   },
   "dependencies": {
     "@parameter1/base-cms-object-path": "^2.5.0",
+    "@parameter1/base-cms-utils": "^2.4.2",
     "apollo-cache-inmemory": "^1.6.6",
     "apollo-client": "^2.6.10",
     "apollo-link-http": "^1.5.17",
